requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#

click==6.6                # via pip-tools
first==2.0.1              # via pip-tools
jsonschema==2.5.1
pip-tools==1.6.5
py==1.4.31                # via pytest
pytest==2.9.2
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0               # via pip-tools, python-dateutil
